% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercer.R
\docType{package}
\name{coercer}
\alias{coercer}
\alias{coercer-package}
\title{The Coercer Package}
\description{
Three coercer functions
}
\section{Details}{

This package contains several functions to convert multiple variables to
either numeric, factors or characters at once: see \code{\link{intonum}},
\code{\link{intofact}}, and \code{\link{intochar}}.
}

\section{Usage}{

intonum(df = data, columns = c(1, 16:20))
}

