% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{get_method2}
\alias{get_method2}
\title{Select a binary method}
\usage{
get_method2(generic, x, y, env = caller_env())
}
\arguments{
\item{generic}{The name of the generic as a string.}

\item{x}{Objects to dispatch on.}

\item{y}{Objects to dispatch on.}

\item{env}{The environment in which to find or define methods.}
}
\description{
This retrieves the method selected by binary dispatch without
calling it. The arguments dispatched on are stored in the
\code{.dispatched} pronoun.
}
