% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intochar_function.R
\name{intochar}
\alias{intochar}
\title{Turn multiple variables into character vectors}
\usage{
intochar(df, columns)
}
\arguments{
\item{df}{Data frame}

\item{columns}{Numeric vector}
}
\value{
New structure of data frame
}
\description{
More detailed description
}
