% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intonum_function.R
\name{intonum}
\alias{intonum}
\title{Turn multiple variables into numeric ones}
\usage{
intonum(df, columns)
}
\arguments{
\item{df}{Data frame}

\item{columns}{Numeric vector}
}
\value{
New structure of data frame
}
\description{
More detailed description
}
