% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-coerce.R
\name{muffle_vec_coerce}
\alias{muffle_vec_coerce}
\title{Muffle vector coercion warnings}
\usage{
muffle_vec_coerce(expr)
}
\arguments{
\item{expr}{Expression to evaluate.}
}
\description{
Evaluate \code{expr} with all vector coercion warnings turned off. Other
warnings are not muffled.
}
\examples{
# Coercing incompatible factors is lossy because it returns a
# character vector that has lost the levels information. Use
# muffle_vec_coerce() to silence this kind of warning:
f1 <- factor("foo")
f2 <- factor("bar")
muffle_vec_coerce(vec_coerce(f1, f2))
}
