% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-coerce.R
\name{vec_coerce}
\alias{vec_coerce}
\title{Vector coercion}
\usage{
vec_coerce(from, to)
}
\arguments{
\item{from}{The vector to coerce. The return value should be a
vector as long as \code{from}.}

\item{to}{The vector containing the type information to determine
the coercion of \code{from}.}
}
\description{
\code{vec_coerce()} is a \link[=dispatch2]{binary generic} meant to power
automatic coercions in tidyverse and r-lib packages. Methods should:
\itemize{
\item Have symmetric coercion behaviour. The resulting type should be
identical on permutation of \code{from} and \code{to}.
\item Not necessarily be fully commutative: \code{vec_coerce()} is expected
to return a vector as long as the first argument \code{from}.
\item Differentiate between empty and non-empty vectors. Most of the
time all the type information should be contained in attributes
and the return type can be determined on empty vectors.

In some cases however the contents of the vector are necessary to
determine the appropriate coercion (e.g. the contents of a
character vector are needed to determine congruence with the
levels of a factor).
}
}
