% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{whichever}
\alias{whichever}
\title{Wildcard type}
\usage{
whichever()
}
\description{
Use \code{whichever()} when the type does not matter.
}
\examples{
def_method2(whichever(), whichever(), fn = function(x, y) "foo")
}
