% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercer.R
\docType{package}
\name{coercer}
\alias{coercer}
\alias{coercer-package}
\title{coercer: A package for coercing multiple variables at once.}
\description{
The coercer package provides three categories of important functions:
intonum, intofact and intochar.
}
\section{Coercer functions}{

The coercer functions ...
}

